RAC-307: Add confirm on stop action
